Refactor System to utilize shared_ptr for better memory management
Replaced raw Property, Client, and Host objects with shared_ptr to enhance memory management and consistency across the System class and related operations. Adjusted method signatures, data structures, and logic accordingly to support the shared_ptr implementation, ensuring safer and more efficient usage of dynamic memory in the application.

+++ Property.cpp
@@ -27,8 +27,8 @@ RentedProperty Property::getRentedProperty() { return this->rentedProperty; }
 
 
 void Property::status() const {
-    std::cout << "Stars: " << this->stars.getStars() << std::endl;
-    std::cout << "Number of avaliations: " << this->stars.amountOfAvaliations() << std::endl;
+    std::cout << "Stars: " << static_cast<int>(this->stars.getStars()) << std::endl;
+    std::cout << "Number of avaliations: " << static_cast<int>(this->stars.amountOfAvaliations()) << std::endl;
     std::cout << "Address: " << this->address << std::endl;
     std::cout << "Owner Name: " << this->ownerName << std::endl;
     std::cout << "Location Type: " << this->locationType << std::endl;
